+++ tests/Fixtures/TaxClassesTest.php
@@ -1,0 +1,44 @@
+<?php
+
+/**
+ *
+ * DISCLAIMER
+ *
+ * Do not edit or add to this file if you wish to upgrade the MultiSafepay plugin
+ * to newer versions in the future. If you wish to customize the plugin for your
+ * needs please document your changes and make backups before you update.
+ *
+ * @category    MultiSafepay
+ * @package     Connect
+ * @author      TechSupport <integration@multisafepay.com>
+ * @copyright   Copyright (c) MultiSafepay, Inc. (https://www.multisafepay.com)
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+ * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+class TaxClassesTest {
+
+    public function getTaxClasses($tax_rate_id) {
+
+        $tax_classes = array(
+            'title' => 'VAT 21',
+            'description' => 'VAT 21 Netherlands',
+            'tax_rule' => array(
+                array(
+                    'tax_rate_id' => $tax_rate_id,
+                    'based' => 'payment',
+                    'priority' => '1',
+                )
+            ),
+        );
+
+        return $tax_classes;
+
+    }
+
+}